Update WRKOBJSQL.SQLRPGLE
Corrected an issue with IBM i V7R2 compatibility. The SYSVIEWS table does not use the same naming convention as the SYSROUTINES table (figures!) so the ORDER BY was failing on V7R2. Apparently this issue is ignored on V7R3 and later, but the code is now modified to handle it on any release.